Add a missing guard clause to application_spec.rb

+++ spec/application_spec.rb
@@ -14,7 +14,7 @@ describe HDOC::Application do
       File.write(@target_file, File.read(@target_file + '-backup'))
       File.delete(@target_file + '-backup')
     else
-      File.delete(@target_file)
+      File.delete(@target_file) if File.exist? @target_file
     end
   end
